+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fix a bug in `sample_u()`
 * Make predict tests opt-in via environment variable
 * Add a vignette comparing the performance of the two MCMC engine options
+* Set the requirement for the nimble package to v1.4.2 or later
 
 ## occumb 1.2.2 (2026/1/9)
 * This patch release includes internal changes for the anticipated JAGS 5.0.0 release and fixes for math rendering issues in the package vignettes.